requirements.txt
# Core dependencies
networkx>=3.0,<4.0
numpy==1.24.3  # Fixed version for compatibility
spacy==3.5.3   # More stable version
thinc==8.1.12  # Specific version for compatibility
pyvis>=0.3.2
fastapi>=0.104.0
uvicorn>=0.24.0
aiohttp>=3.9.0
urllib3<2.0.0  # Compatible with LibreSSL

# Machine Learning (optional - will be installed if possible)
torch>=2.0.0
transformers>=4.30.0

# Web server
httptools>=0.6.0
websockets>=12.0
uvloop>=0.19.0; sys_platform != 'win32'

# Utilities
pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# NLP and embeddings
nltk>=3.8.0
scikit-learn>=1.3.0
tokenizers>=0.13.0
gensim>=4.3.0 

pyttsx3>=2.90
